Time: 71 ms (97.99%), Space: 32.3 MB (40.04%) - LeetHub

+++ 1465-maximum-area-of-a-piece-of-cake-after-horizontal-and-vertical-cuts/1465-maximum-area-of-a-piece-of-cake-after-horizontal-and-vertical-cuts.cpp
@@ -21,7 +21,7 @@ public:
             maxc=max(maxc,ve[i]-ve[i-1]);
         }
         
-        long long int ans=(1LL*maxr*maxc)%MOD;
+        int ans=(1LL*maxr*maxc)%MOD;
         return ans;
     }
 };
